+++ rust/operator-binary/src/controller/validate.rs
@@ -233,6 +233,8 @@ pub fn validate_cluster(
             kerberos_secret_class,
         },
         role_group_configs,
+        dereferenced_objects.role_listener,
+        dereferenced_objects.existing_discovery_config_map,
     ))
 }
 
@@ -296,6 +298,8 @@ mod tests {
             DereferencedObjects {
                 s3_connection_spec: None,
                 hive_opa_config: None,
+                role_listener: None,
+                existing_discovery_config_map: None,
             },
         )
     }