+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 - BREAKING: The `metastore` role is now required by the CRD; a HiveCluster without it was
   previously accepted by the API server but failed reconciliation ([#731]).
 - Bump stackable-operator to 0.114.0 ([#735]).
+- The reconciler now applies resources and derives the cluster status in discrete
+  apply and update_status steps ([#737]).
 - All product containers now run with `securityContext.runAsNonRoot` set to `true` to improve security ([#741]).
 
 ### Fixed
@@ -27,6 +29,7 @@ All notable changes to this project will be documented in this file.
 [#731]: https://github.com/stackabletech/hive-operator/pull/731
 [#735]: https://github.com/stackabletech/hive-operator/pull/735
 [#736]: https://github.com/stackabletech/hive-operator/pull/736
+[#737]: https://github.com/stackabletech/hive-operator/pull/737
 [#741]: https://github.com/stackabletech/hive-operator/pull/741
 
 ## [26.7.0] - 2026-07-21